Add ForgotPasswordViewModel
This commit introduces the `ForgotPasswordViewModel` class, which handles the logic for the forgot password feature.

Key changes:
- Created `ForgotPasswordViewModel` class that extends `ViewModel`.
- Added `UserRepository` dependency for interacting with user data.
- Implemented `resetPassword` method to initiate the password reset process.
- Added LiveData objects to observe the reset status, errors, and validation exceptions.
- Included `CompositeDisposable` to manage RxJava subscriptions.
