+++ app/src/main/java/com/hminq/quizlett/ui/home/HomeFragment.java
@@ -1,0 +1,51 @@
+package com.hminq.quizlett.ui.home;
+
+import android.os.Bundle;
+
+import androidx.annotation.NonNull;
+import androidx.annotation.Nullable;
+import androidx.fragment.app.Fragment;
+import androidx.lifecycle.ViewModelProvider;
+
+import android.util.Log;
+import android.view.LayoutInflater;
+import android.view.View;
+import android.view.ViewGroup;
+
+import com.hminq.quizlett.R;
+import com.hminq.quizlett.ui.SharedViewModel;
+
+import dagger.hilt.android.AndroidEntryPoint;
+
+@AndroidEntryPoint
+public class HomeFragment extends Fragment {
+    private static final String TAG = "FRAGMENT_HOME";
+    private SharedViewModel sharedViewModel;
+
+    public HomeFragment() {}
+
+    @Override
+    public void onCreate(Bundle savedInstanceState) {
+        super.onCreate(savedInstanceState);
+
+        sharedViewModel = new ViewModelProvider(requireActivity()).get(SharedViewModel.class);
+    }
+
+    @Override
+    public View onCreateView(LayoutInflater inflater, ViewGroup container,
+                             Bundle savedInstanceState) {
+        // Inflate the layout for this fragment
+        return inflater.inflate(R.layout.fragment_home, container, false);
+    }
+
+    @Override
+    public void onViewCreated(@NonNull View view, @Nullable Bundle savedInstanceState) {
+        super.onViewCreated(view, savedInstanceState);
+
+        sharedViewModel.getCurrentUserLiveData().observe(requireActivity(), currentUser -> {
+            if (currentUser != null) {
+                Log.d(TAG, "Current user: " + currentUser.getEmail());
+            }
+        });
+    }
+}
